fix(server): bring back port to uin16. It will never go this far

+++ server/src/main.cpp
@@ -9,11 +9,11 @@
 #include "scenes.hpp"
 
 int main(int ac, char **av) {
-    uint32_t port = 6767;
+    uint16_t port = 6767;
     std::string protocol = "UDP";
 
     if (ac > 1) {
-        port = static_cast<uint32_t>(std::stoi(av[1]));
+        port = static_cast<uint16_t>(std::stoi(av[1]));
     }
     if (ac > 2) {
         protocol = av[2];
